refactor core/matrix
