+++ core/src/matrix/mod.rs
@@ -1,4 +1,6 @@
 pub mod hash;
 pub mod traits;
+pub mod utils;
 
-pub use traits::{MatrixMulExt, expand_a_from_rho};
+pub use traits::MatrixMulExt;
+pub use utils::expand_a_from_rho;
